+++ apps/arceos/ivc_publisher/src/main.rs
@@ -15,19 +15,25 @@ fn main() {}
 
 #[cfg(feature = "arceos")]
 mod publisher {
-    use core::{cell::UnsafeCell, result::Result::Err};
+    use core::{cell::UnsafeCell, result::Result::Err, sync::atomic::AtomicU64};
 
     use ax_std::{
-        os::arceos::modules::ax_hal::mem::{PhysAddr, VirtAddr, virt_to_phys},
+        os::arceos::modules::ax_hal::{
+            irq,
+            mem::{PhysAddr, VirtAddr, virt_to_phys},
+        },
         println,
     };
     use axhvc::ivc::{self, IvcGuestPhysAddr};
-    use axivc::{IVC_SLOT_PAYLOAD_SIZE, IvcRegion};
+    use axivc::{IVC_SLOT_PAYLOAD_SIZE, IvcPeerEventWaiter, IvcRegion, record_peer_event};
 
     const PUBLISH_COUNT: u64 = 5;
+    static NOTIFY_IRQ_COUNT: AtomicU64 = AtomicU64::new(0);
+
     mod demo_config {
         pub const CHANNEL_KEY: usize = 0x4956_4301;
         pub const CHANNEL_SIZE: usize = 4096;
+        pub const NOTIFY_IRQ: Option<usize> = Some(60);
         pub const PUBLISHER_VM_ID: usize = 1;
         pub const SUBSCRIBER_VM_ID: usize = 2;
     }
@@ -61,35 +67,46 @@ mod publisher {
             return;
         };
         region.initialize(demo_config::PUBLISHER_VM_ID, demo_config::CHANNEL_KEY);
-        run_request_ack_demo(region);
+        let waiter = IvcPeerEventWaiter::new(register_notify_irq(), &NOTIFY_IRQ_COUNT);
+        run_request_ack_demo(region, &waiter);
     }
 
-    fn run_request_ack_demo(region: &'static IvcRegion) {
+    fn run_request_ack_demo(region: &'static IvcRegion, waiter: &IvcPeerEventWaiter<'_>) {
         let mut ack_payload = [0u8; IVC_SLOT_PAYLOAD_SIZE];
         let mut subscriber_ready = false;
         for sequence in 1..=PUBLISH_COUNT {
-            send_request_when_ready(region, sequence, subscriber_ready);
-            wait_for_ack(region, sequence, &mut ack_payload);
+            send_request_when_ready(region, sequence, subscriber_ready, waiter);
+            wait_for_ack(region, sequence, &mut ack_payload, waiter);
             subscriber_ready = true;
         }
     }
 
-    fn send_request_when_ready(region: &'static IvcRegion, sequence: u64, subscriber_ready: bool) {
+    fn send_request_when_ready(
+        region: &'static IvcRegion,
+        sequence: u64,
+        subscriber_ready: bool,
+        waiter: &IvcPeerEventWaiter<'_>,
+    ) {
         loop {
             match region.send_request(sequence, b"hello from arceos publisher") {
                 Ok(()) => {
+                    println!("ivc send seq={sequence}");
                     if subscriber_ready {
                         notify_subscriber();
                     }
-                    println!("ivc send seq={sequence}");
                     return;
                 }
-                Err(_) => wait_for_subscriber_poll(),
+                Err(_) => waiter.wait_for_peer_event(),
             }
         }
     }
 
-    fn wait_for_ack(region: &'static IvcRegion, sequence: u64, payload: &mut [u8]) {
+    fn wait_for_ack(
+        region: &'static IvcRegion,
+        sequence: u64,
+        payload: &mut [u8],
+        waiter: &IvcPeerEventWaiter<'_>,
+    ) {
         loop {
             match region.try_recv_ack(payload) {
                 Ok(Some(message)) if message.sequence() == sequence => {
@@ -98,7 +115,7 @@ mod publisher {
                     println!("ivc ack seq={} msg={text}", message.sequence());
                     return;
                 }
-                Ok(Some(_)) | Ok(None) => wait_for_subscriber_poll(),
+                Ok(Some(_)) | Ok(None) => waiter.wait_for_peer_event(),
                 Err(err) => {
                     println!("ivc publish failed: recv ack error {err:?}");
                     return;
@@ -117,9 +134,38 @@ mod publisher {
         }
     }
 
-    fn wait_for_subscriber_poll() {
-        for _ in 0..100_000 {
-            core::hint::spin_loop();
+    fn register_notify_irq() -> bool {
+        let Some(raw_irq) = demo_config::NOTIFY_IRQ else {
+            return false;
+        };
+        match notify_irq_id(raw_irq)
+            .and_then(|irq_id| irq::request_shared_irq(irq_id, notify_irq_handler).map(|_| irq_id))
+        {
+            Ok(irq_id) => {
+                println!("ivc notify irq enabled irq={irq_id:?}");
+                true
+            }
+            Err(err) => {
+                println!("ivc notify irq disabled raw={raw_irq} err={err:?}");
+                false
+            }
+        }
+    }
+
+    fn notify_irq_handler(_ctx: irq::IrqContext) -> irq::IrqReturn {
+        record_peer_event(&NOTIFY_IRQ_COUNT);
+        irq::IrqReturn::Handled
+    }
+
+    fn notify_irq_id(raw_irq: usize) -> Result<irq::IrqId, irq::IrqError> {
+        #[cfg(target_arch = "aarch64")]
+        {
+            let gsi = u32::try_from(raw_irq).map_err(|_| irq::IrqError::InvalidIrq)?;
+            irq::resolve_irq_source(irq::IrqSource::AcpiGsi(gsi))
+        }
+        #[cfg(not(target_arch = "aarch64"))]
+        {
+            irq::try_legacy_irq(raw_irq)
         }
     }
