+++ virtualization/axdevice/src/lib.rs
@@ -77,7 +77,7 @@ pub use loongarch_pch_pic::{
     PchPicOutputPort, PchPicOutputPortKey,
 };
 pub use model::{DeviceFirmwareProperty, DeviceFirmwareSpec, DeviceModel};
-pub use range_alloc::{GuestRangeAllocator, GuestRangeAllocatorKey, GuestRangePool};
+pub use range_alloc::{GuestRangeAllocator, GuestRangeAllocatorKey, GuestRangePool, IvcNotifyIrqKey};
 pub use registration::{DeviceBundle, DeviceLifecycle, DeviceRegistration, PollableDeviceOps};
 pub use resources::{
     DevicePlanRequest, DeviceRequirement, DeviceRequirements, MsiResourceRequest,
