+++ virtualization/axivc/.gitignore
@@ -1,0 +1,12 @@
+/target
+/.vscode
+.DS_Store
+Cargo.lock
+
+# Test results (generated by shared test framework)
+/test-results/
+/test_repos/
+*.log
+
+# Downloaded test framework
+/scripts/.axci/
